New module: filesystem.ts 1. moved checkFirstUse from save.ts to filesystem.ts 2. moved functions from fetch.ts to filesystem.ts 3. renamed dirname.ts to filesystem.ts
