+++ src/lib/export.ts
@@ -1,10 +1,10 @@
 import chalk from 'chalk';
 import Watch from 'node-watch';
-import __dirname from '../api/dirname.js';
 import { readdir } from 'node:fs/promises';
 // @ts-ignore
 import selectFolder from 'win-select-folder';
-import { getFile, getImage } from '../api/fetch.js';
+import { __dirname } from '../api/filesystem.js';
+import { getFile, getImage } from '../api/filesystem.js';
 import { fileExists, listFiles, updateFile, uploadFile } from '../api/supabase.js';
 
 type Images = {
